backend/requirements.txt
flask==2.3.3
pyjwt==2.8.0
python-dotenv==1.0.0 